backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flasgger==0.9.7.1
pandas==2.1.4
pytesseract==0.3.10
Pillow==10.1.0
rapidfuzz==3.5.2
requests==2.31.0
PyPDF2==3.0.1
easyocr==1.7.0
python-dotenv==1.0.0
